# homelike-service/src/swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:30021/api",
      "description": "Docker"
    },
    {
      "url": "http://localhost:30020/api",
      "description": "Local"
    }
  ],
  "info": {
    "version": "v1.0",
    "title": "homelike-app",
    "description": "Node.js + Typescrypt + MongoDB Real Estate API as a proof of concept."
  },
  "paths": {
...
    "/users/authenticate": {
      "post": {
        "summary": "authenticate",
        "description": "Login a user",
        "operationId": "authenticate",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "name": {
                      "type": "string"
                    },
                    "first_name": {
                      "type": "string"
                    },
                    "last_name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "example": {
                  "email": "jancaldron@gmail.com",
                  "password": "123456"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}